0.6.12.3: 	NOTE: This is a slightly screwed-up checkin, which won't 		quite build as is, because the checked-in changes from 		the Alpha port expect various src/code/*types.lisp 		files (e.g. src/code/x86-linux-types.lisp) which 		Daniel Barlow created by hand once and for all. Those 		are in my working directory but I didn't check them in 		because almost immediately, probably 0.6.12.4, I 		expect to rearrange things so that the build 		process generates a types file on every run of 		make.sh. So since CVS add/delete is a little mixed up, 		I thought I'd just skip it, even though it means that 		this version isn't quite buildable unless you grab 		those files from DB's patch. 	merged Daniel Barlow's port to the Alpha CPU sbcl-devel 		2001-05-04 	moved CIRCULAR-LIST-P to early-extensions.lisp (since it 		logically belongs there, and DB's new definition, 		unlike the old one, doesn't have a physical dependence 		on IGNORE-ERRORS which prevents it from being there) 	merged boot-extensions.lisp, early-extensions.lisp, and 		late-extensions.lisp, since there's no longer any 		distinction between the files 	deleted CVS files from sbcl-alpha-extra-files.tar, since 		they were probably an oversight, I think 	also src/compiler/alpha/z050.pdf:-| 	The patch change 			-  (/hexstr arguments) 			+  #!+sb-show (/hexstr arguments) 		shouldn't be necessary because /SHOW, like all the 		/FOO macros, is already conditional on SB-SHOW. So I 		undid this change (and added some comments in 		src/code/show.lisp explaining the intended behavior 		of /FOO macros). 	merged stat_wrapper.c into the relatively new, related 		wrap.c file; deleted stat_wrapper.h since stat wrappers 		seem never intended to be used by C code anyway 	renamed build-tools/ to tools-for-build/ in the interest of 		unambiguous (even painfully unambiguous) 		globally-visible names 	tweaked tests/alien.impure.lisp so that the test uses a 		non-GENCGC-dependent variable (instead of having to 		have the test conditional on #+GENCGC).

+++ src/runtime/arch.h
@@ -29,7 +29,7 @@ extern unsigned long arch_install_breakpoint(void *pc);
 extern void arch_remove_breakpoint(void *pc, unsigned long orig_inst);
 extern void arch_install_interrupt_handlers(void);
 extern void arch_do_displaced_inst(os_context_t *context,
-				   unsigned long orig_inst);
+				   unsigned int orig_inst);
 extern lispobj funcall0(lispobj function);
 extern lispobj funcall1(lispobj function, lispobj arg0);
 extern lispobj funcall2(lispobj function, lispobj arg0, lispobj arg1);
